Clean-up of the build.gradle files and refactoring of the modules so that FES module only depends on vitrivr-engine-core
