+++ gradle.properties
@@ -17,7 +17,7 @@ version_kotlinx_serialization=1.6.2
 version_kotlinx_datetime=0.4.1
 version_kotlinlogging = 5.1.0
 version_log4j2=2.20.0
-version_moshi=1.15.1
+version_metadataextractor=2.19.0
 version_okhttp3=4.12.0
 version_picnic=0.7.0
 version_protobuf=3.25.1
