+++ vitrivr-engine-core/build.gradle
@@ -7,9 +7,9 @@ plugins {
 dependencies {
     /** JOML dependencies for 3D mesh support. */
     implementation group: 'org.joml', name: 'joml', version: version_joml
+
     /** dependencies for exif metadata extraction. */
-    implementation 'com.drewnoakes:metadata-extractor:2.19.0'
-    implementation 'com.google.code.gson:gson:2.8.9'
+    implementation group: 'com.drewnoakes', name: 'metadata-extractor', version: version_metadataextractor
     implementation group: 'io.javalin.community.openapi', name: 'javalin-openapi-plugin', version: version_javalinopenapi
 }
